Add a "ship" and let the player move around.
This includes some tile / palette data in a separate file and some code
for DMA'ing that data into VRAM. This is used to render a "ship" (white
circle) that the player can move around with the D-Pad.

Added a whole bunch of named registers (& documentation for them) to
registers.asm.